+++ src/test/app/agent/blog_agent_test.py
@@ -1,6 +1,10 @@
+import os
+from unittest import mock
+
 import logging.config
 import unittest
 
+from aideas.app.agent.translation.translator import Translator
 from test.app.agent.test_blog_agent import TestBlogAgent
 from test.app.test_functions import init_logging, load_agent_config, get_run_context, run_agent
 from aideas.app.agent.agent_name import AgentName
@@ -18,9 +22,26 @@ class BlogAgentTest(unittest.TestCase):
             run_context.get_app_config().to_dict(),
             load_agent_config(agent_name, False))
 
-        result = run_agent(agent, run_context)
+        with mock.patch.object(Translator, 'translate_file_path') as translate_file_path:
+            with mock.patch.object(Translator, 'translate') as translate:
 
-        self.assertTrue(result.is_successful())
+                translate.side_effect = lambda text, from_lang_code, to_lang_code: text
+
+                files_to_delete = []
+
+                def mock_translate_file_path(file_path, _, to_lang_code):
+                    name, ext = os.path.splitext(os.path.basename(file_path))
+                    translated_file_path = os.path.join(os.path.dirname(file_path), f"{name}.{to_lang_code}{ext}")
+                    files_to_delete.append(translated_file_path)
+                    return translated_file_path
+
+                translate_file_path.side_effect = mock_translate_file_path
+
+                try:
+                    result = run_agent(agent, run_context)
+                    self.assertTrue(result.is_successful())
+                finally:
+                    [os.remove(f) for f in files_to_delete if os.path.exists(f)]
 
 
 if __name__ == '__main__':